requirements.txt
torchvision
braceexpand
webdataset
tqdm
wandb
einops
einops-exts
transformers>=4.36.1
torch>=2.0.1
pillow
open_clip_torch>=2.16.0
sentencepiece
open_flamingo
jupyter
datasets
evaluate
peft
accelerate
trl
huggingface_hub
bitsandbytes
deeplake